ERR: Problem with vtkClipPolyData reported by Paul Hsieh: pahsieh@usgs.gov outCD was being used by both outputs. A separate outClippedCD was created. ---------------
I am running into a problem with vtkClipPolyData when the
dataset contains cell scalars. The filter does not seem to
work correctly with GenerateClippedOutputOn, but
works fine with GenerateClippedOutputOff.

This problem is illustrated by the attached sample code.
A structured points data set is created with cell scalars,
and is then passed through a geometry filter to obtain
a poly dataset for clipping.

Clipping is done twice on the poly dataset, the first time
with GenerateClippedOutputOn, and the second time with
GenerateClippedOutputOff. The resulting actors are different.

This problem occurs in both vtk 2.3 and vtk 3.1 ("official"
release). The problem also occurs when the clipped dataset is
a structured grid that has been passed through a geometry filter,
or a poly data created from scratch.

Thanks for any help and best regards,
-- Paul