Generalized the "oblique angles" concept for parallel projections into the "ViewShear" vector.  The view frustum can be sheared in x and y arbitarily now.  The old ObliqueAngles parameter now sets the view shear.
Note: there's a problem where interactive rendering with ViewShear
turned on causes the computed near and far clipping planes to be
incorrectly set, possibly cutting off part of the image.  It isn't a
problem with the ViewShear code, as far as I can tell --- it's
somewhere else.

Also added a clarification to the "ParallelScale" documentation.  I
think this parameter is incorrectly named: it scales the viewport
height, not the size object as I would expect a "scale" to do.