Some corrections to color space conversions.
STYLE: The conversion between RGB and XYZ used the sRGB standard, but the sRGB is a non-linear color space.  Thus, it was invalid to do OpenGL color computations in it.  It also made colors look overly bright because most systems do some gamma correction of their own.

STYLE: The conversion from RGB to XYZ multiplied everything by a factor of 100.  I could not find anything in the liturature to suggest this was common practice.  I removed this multiplication.