ENH: Add concept of Resolution to the pipeline.
You can now request data at a particular resolution. Do so by having the
sink call SetUpdateResolution(0.0->1.0) in the same way that it calls
SetUpdateExtent to ask for a particular piece. Similar to the way time
works, this new request goes up the pipeline to the source which can respect
the request or not, and returns an answer in the form of a DATA_RESOLUTION
key in its produced DataObject.