KWSys: Fix SharedForward on Cygwin without -mwin32
When building on Cygwin without -mwin32, the _WIN32 macro may not be
defined.  SharedForward must still set the PATH environment variable to
ensure runtime dependencies are found.

The 'ldd' wrapping feature uses 'cygcheck' for now since a real ldd tool
is not available in Cygwin 1.5.  We can change to use the real ldd when
we choose to stop supporting legacy Cygwin and require 1.7.