BUG: Fixed the issue of clipping near by objects. Also fixed the issue with missing tiles when zoomed in.
Fixed the issue with huge offset between line drawing and polygon drawing.\n\n The reason was Paraview adding ZBuffer offset globaly during processing of modules.
Fixed the issue with clipping range resetting.
Found a workaround for the issue of missing tiles when zoomed in too close to the planet.
Saving and restoring the depth offset state.
Fixed the issue with actors not added to the representation. The issue is somewhat related to rendering / update.

Remove debug messages.