BUG: Temporary fix for memory use in text rendering.
This is a temporary fix for memory use in the Qt label render strategy.
The strategy caches all unique strings, and in the case of charts this
could be a large number of strings (hundreds of MBs). Until the label
render strategy is fixed the approach taken here is to delete the object
and create another one every 300 renders with text.