COMP: Revert "Add vtkNew<> to create objects locally"
The syntax

  vtkSmartPointer<vtkBaseClass> b = vtkNew<vtkDerivedClass>();

does not work as implemented.  C++98 12.2/2 allows the compiler to
copy-construct an extra vtkNew<> temporary to which to bind the
reference of b's constructor.  Since vtkNew<> does not provide copy
semantics (and cannot by design) this fails on compilers that choose to
copy the temporary unnecessarily.

I'm reverting this feature until we have time to investigate further.