ENH: Refactored label generation, set text labels.
Added the ability to set arbitrary labels at arbitrary points on the
axis. This is accomplished by setting the tick labels and positions
variables on the axis. This allows for things such as string
enumerations to be plotted if put through the correct filter. Also
refactored the way that the labels are generated (should be generated
less, and less susceptible to rounding issues around min/max).