Fixes to vtkDataSetGradient for 0 1 and 2D cells.
When summing the CQS vectors to create a gradient, the values are weighted
by the volume of the cell.  Fine, but that only makes sense for 3D cells.  2D
cells technically have no volume.  Their volumes were set to 0, and when
dividing by that you clearly got bad results.

You need to weight by the area for 2D cells.  Rather than compute volume,
the precompute now computes "size" which is volume, area, and line for 3D,
2D, and 1D cells, respectively.  Size is undefined for 0D cells.  It is set
to 1 simply so that it creates a no-op in the weighting.

Also added CQS calculations for 0D and 1D cells.