ENH: Rework vtkWrapJava and vtkWrapTcl to use vtkParseType.h
All of the hexidecimal constants in vtkWrap/ParseJava and vtkWrapTcl
have been replaced with constants defined in vtkParseType.h.  The
semantics of these files is the same as before.  I checked the
generated code for all VTK classes to make sure that it is the same
as it was before this change.