ENH: Move core python support classes to Wrapping/Python
Organizationally, it makes more sense for the python utility classes
to be placed in Wrapping/Python than in Common.  This means adding a
new library called vtkPythonCore that represents the core python
support for VTK, as compared to VTK classes for individual kits like
Common/vtkPythonCommand.cxx and Rendering/vtkTkRenderWidgetPython.cxx.