ENH: More vtkParse enhancements, better method signatures
The main change here is that ArgClasses and ReturnClass are set for
all types, not just VTK_PARSE_VTK_OBJECT.  The main benefit of this
is that even if the type is VTK_PARSE_UNKNOWN, a string that gives
the class name will be available, and the string will show any scoping
or template-types for the class.

As well, the code for building type signatures has been cleaned up,
and enums are parsed completely but not yet stored in the output.