BUG: Somtimes vtkParse would get method comments out-of-order
The method comments are collected by lex, which is always ahead of
yacc by one or more tokens.  Sometimes lex would collect the comment
for the next method before output_method() had been called for the
current method, and the current method would get the comment for the
next method.  This change fixes the problem by having yacc collect
the comment earlier.