BUG: add pythonic signatures back into __doc__ strings
Recently I changed the formatting of the docstrings to handle
some doxygen and html tags and to break lines at 70 chars.
As part of this clean-up, I removed the pythonic signatures from
the docstrings and just left the C++ signatures.  However,
it seems that Vistrails and maybe other projects rely on the
pythonic signatures for introspection.