COMP: force exclude vtkgl.h from hierarchy file
For some reason, Win32 builds could not find this file in the
CMAKE_CURRENT_BINARY_DIR.  Maybe because the BINARY_DIR was a
relative path, or maybe the dependencies were wrong and the file
wasn't built.  Oddly, vtkTypeInt32Array.h and friends are also
in CMAKE_CURRENT_BINARY_DIR (for Common) and they work fine.