ENH: Add VTK_IGNORE_BTX option to ignore BTX/ETX markers.
This commit adds more automatic hinting for the wrappers,
enough so that the wrappers will compile even when the BTX
markers are ignored.  The vtkWrapHierarchy tool now includes
typedefs and enums, as well as cmake flags ABSTRACT, WRAP_EXCLUDE,
and WRAP_SPECIAL.  The wrappers can inspect these values for
any class in the hierarchy.