Replaced Delete with FastDelete (when possible) for optimization.
When working with lots of objects, for example in large AMR
datasets, garbage collection becomes a serious bottleneck. Changes
in this commits are the result of profiling the processing of a
large AMR dataset. Note that it is not always appropriate to use
FastDelete instead of Delete. Use it only when you are sure that
it will not cause the object to be deleted. For example, right
after you pass it to a Set function that increases the argument's
reference count.