Improved vtkCellDataToPointData for vtkUnstructuredGrid datasets.
vtkCellDataToPointData creates cell links when it traverses the datasets
which is an expensive operation for vtkUnstructuredGrid datasets. The
improved vtkCellDataToPointData behaves identically for all other types
of datasets as befores with the only differences working with
vtkUnstructuredGrid.  For a vtkUnstructuredGrid, it traverses the cell
array directly and keeps two auxiliary arrays holding the accumulated
quantity in question and the number of cells sharing each point. After
the traverse is done, it dose an extra division pass to get the average
of the tracked quantity which returns exact the same results as before.