Add vtkMatrixMathFilter which provides tensor operation for point/cell data.
vtkMatrixMathFilter currently provides four types of tensor operations for
point/cell associated tensor data. The operations are determinant, inverse,
eigenvalue, eigenvector. The actual computation is delegated to vtkMath as
much as possible for consistent behavior of the system. The only
exception is inverse because the one in vtkMath does not check for
singularity of the input matrix which might cause divide by zero cases. To
ensure consistent behavior, the original code is copied into position with
the necessary added.