Delay Git hooks setup until first commit
Install a small .git/hooks/pre-commit script in the source tree at CMake
configuration time instead of telling the user to do it.  When the user
first attempts to commit this bootstrapping hook will trigger and print
instructions to install the real hooks.  This approach allows us to
require local hooks for commit but not for those who just want to build.