COMP: Fix Java dependencies wrt Hierarchy build.
The .java files were all trying to build up-front and were causing
the Hierarchy files to build out of order.  Making them depend on
their kit library seems to fix the problem (it is not a true
dependency, but it helps to make things occur in the correct
order).  Also, the header files are now direct dependencies
of the hierarchy file that is created from them.