Fixed BUG #11016. Don't bother not copying input scalars.
Refer to BUG #8494 and BUG #11016. I cannot see any reason why one would
want to turn CopyScalars Off. My understanding is that this was done to
avoid copying of "ClipDataSetScalars" to the output when
this->GenerateClipScalars is false. But, if GenerateClipScalars is false,
then "ClipDataSetScalars" is not added as scalars to the input at all
(refer to code above) so it's a non-issue. Leaving CopyScalars untouched
i.e. ON avoids dropping of arrays (#8484) as well as segfaults (#11016).