ENH: New test for line plot mouse interaction.
Thanks to Karthik for explaining the mouse interaction testing, and Bill
Lorensen for showing me the vtkTestingObjectFactory code. This is the
first VTK charts test with mouse interaction tests. It manipulates the
chart in several ways, and then records an image of the end result after
panning, zooming, selecting and panning.