BUG: 11208. Fix vtkOpenGLExtensionManager crash on VS2003.
ReadOpenGLExtensions() calls this->RenderWindow->Render() on Win32 machines if Render() has not previously been called. This caused a crash, because Render() calls vtkOpenGLRenderWindow::OpenGLInit(), which was in turn calling vtkOpenGLRenderWindow::SetExtensionManager(0) and then creating a new ExtensionsManager.  The crash only seemed to occur on VS2003. See bug report for more details.