BUG: Fix segfault on table switchting.
Thanks to Eric Monson for pointing out the bug. Added a new member
function to set the visibility for all columns in a parallel coordinate
plot. Used this in the table update function of PlotParallelCoordinates
to reset the visibility before building up the list.

This should also provide a reasonable way to change the visibility of
all columns in a plot with one call.