Leave calling convention off typedef for now
Changing the calling convention in the C++ headers caused
mismatches between the C++ libs and the python and tcl
wrappers, which assume cdecl. This led to failing tests
on the dashboards that test these methods.

For now, simply put the calling convention back to the
way it was before (do not specify it), but keep the typedefs.

More work is required to be able to specify the calling
convention in the typedefs and get all the wrapped languages
to play nicely with them. For now, at least I can have
the typedefs in here.