ENH: Created method for setting tooltip info outside of LocatePointInPlots
Made a new method for setting tooltip info which is split out from
LocatePointInPlots so that subclasses can override this method for
new types of tooltips without having to reimplement so much code.
Tried to make it very general by passing all of the (what I saw
as) useful variables at that point in LocatePointInPlots().