ENH: Added some new 2D histogram chart and items.
This new chart type takes a vtkImageData as input, and produces a plot
using another vtkImageData. The axes and transform are all handled by
the chart. I still need to add an additional item to display the color
mapping and scale. There are a few layout issues to solve too, but all
fo the basics are now working.

Change-Id: Icf8170e89d720fd6381c8ed9c420ae0c2a023c6d