ENH: Create textures from vtkImageData.
This function is still pretty raw, and will only work as expected when
used with an OpenGL 2 card. It is only called in one of the DrawImage
functions right now, and will either be generalized or removed. It is
also hardwired right now for drawing images the way I need them drawn.

Change-Id: Icf3a7160edadb47e2f857430ba7f9ae82a30e2d0