Fix mpi test executable path for Makefile generators
Commit 892cc8e5 (when mpi based tests are run on windows, the mpiexe
path-to-exe does not include the debug/release part of the path and
hence these tests do not run - this fix should resolve it, 2010-03-19)
hard-coded the configuration name into the path to executables
referenced in MPI test command lines.  Do this only for generators that
have multiple configuration types (VS IDE, Xcode).  For Makefile
generators the executable path does not include the configuration type.

This fixes tests that pass the executable location to mpirun when CTest
is invoked with an explicit configuration, as in "ctest -C Debug".