Properly setting the number of pieces and whole extent.
In certain cases the writer didn't have the correct
number of pieces and the whole extent for the
vtkExtentTranslator but this has now been fixed by
properly getting it from the pipeline information
for the whole extent and the number of pieces
from the member variable.

Also got rid of the ExtentTranslator member variable
and setter/getter functions since it's not used
and shouldn't be.