ENH: Added circle packing layout classes to VTK.
The vtkCirclePackLayout filter is similar to the area layout
filter in Infovis and produces a triple for each vertex in
an input vtkTree.  The triple represents the (X, Y, Radius)
of the circle for that vertex.  There is a concrete implementation
of the vtkCirclePackLayoutStrategy class for a "front chain"
based circle packing layout.  There is also a regression test
for the layout classes and a poly data mapper class for the
packed circle layout.