Fixing bug where string array lookup fails
vtkStringArray's LookupValue() function uses vtkSortDataArray to
sort values before looking up matching indices. Due to what seems to
be a bug or inefficiency in that function, it can cause long run times
and function stacks length 3000+ when sorting a vtkStringArray with
many repeated values of size 100000+. Switching to STL sort on a map
of pairs decreses runtime from hours to seconds.

Change-Id: I3ddae4dd257e15fa351ed70b9bf44cf11b558ff9