ENH: Make vtkScalarsToColors into a concrete class.
This patch enhances vtkScalarsToColors so that it can be used as a
simple way to do scalar range mapping.  It is meant to be used
when you have RGB scalars and want to map them to "unsigned char"
RGB or RGBA colors, or if you have single-component scalars and
simply want to map them to greyscale.

Basically, it maps scalars directly to colors without the use of a
CTF or a lookup table, all it does is rescale the values.  It also
does the job faster than vtkLookupTable and vtkColorTransferFunction.

Change-Id: I24fa8a1545a969c2ee5202bd8eca76b52842fbf4