BUGFIX: Correct vtkAMRBox::GetCellIndex
Fixed issue with computing the cell id which
was assuming the AMR box dimensions were node-based.
More asserts were added for debugging purposes elsewhere.
Automatic detection of the dimension is currently broken
in the case where we have 2-D or 3-D grids that consist of
a single layer of cells.