Minor fix to the vtkXMLPStructuredDataWriter.
Below is the comment that I added to
vtkXMLPStructuredDataWriter::RequestUpdateExtent in this commit.
It explains the change in detail:

The code below asks for an extent based on the number of pieces and
the first piece. This is mainly for the sake of other filters/writers
that may internally use this reader. Those writers usually delegate
RequestUpdateExtent() to the internal write and expect it to do the
right thing. Before this change, vtkXMLPStructuredDataWriter did not
bother setting the update extent because it is taken care of by the
vtkXMLStructuredDataWriter during WritePiece() (see vtkXMLPDataWriter).
This is fine when vtkXMLPStructuredDataWriter's input is connected
to the actual pipeline but causes problems when it is not, which happens
in the situation described above. Here I picked the StartPiece as
the default. This will not effect the behavior when writing multiple
pieces because that does its own RequestUpdateExtent with the right
piece information.