BUGFIX:Initialize internal reader & indexing
This commit corrects issues due to erroneous
initialization of the internal reader. Specifically,
although, the Internal->FileName was set, the
Internal->BoundaryFileName, Internal->MajorFileName,
Internal->HierarchyFileName and Internal->DirectoryName
were not being set properly.
Furthermore, a minor issue with indexing that was causing
vtkAMREnzoParticlesReader::FindBlockIndex to return false is fixed
-- blocks are numbered from 1.