ENH: Add SelectPoints() functionality to vtkPlotBar
The derived function SelectPoints() is added to vtkPlotBar. When
called, this sets the ivar vtkIdTypeArray* Selection to point to an
object with values corresponding to the indices of selected bars,
and returns a bool indicating whether Selection's referent contains
any values.

SelectPoints() also has the side effect of triggering Paint() to
highlight the selected bars.

Following the behavior of vtkPlotPoints::SelectPoints(),
vtkPlotBar::SelectPoints() will always allocate Selection, regardless
of the return value of the function.

vtkPlotBar::SelectPoints() does not support multiple bar segments. If
called on an vtkPlotBar instance that contains more than 1 bar
segments, the function returns false and does not highlight any bars.
Future support of multiple bar segments may be difficult, as the
Selection array and associated API does not provide for
distinguishing particular segments.

Change-Id: Id8442ccb80320957415a90965438561d940aac50