Ask MPI whether it's been initialized rather than trying to track it ourselves
Calling MPI_Init multiple times in an application is an error.  This comes up
a lot when trying to run Python scripts that use VTK MPI components.  A while
back I wrote code to try to trap this by tracking (within VTK) whether we had
already initialized MPI.  This turns out to be brittle.  I'm backing that code
out and replacing it with a call to MPI_Initialized() -- we'll just ask MPI
itself whether it's up and running.