Mark constructors that take array pointers as explicit.
Any single-argument constructor that is not marked as explicit
can be used for automatic argument conversion by C++.  In this
case, it meant that vtkChart::SetSize(const vtkRectf&) could be
called as SetSize(float *), which I do not think was intended.
I believe that explicit conversion SetSize(vtkRectf(float *))
was the intent, and the "explicit" keyword enforces this.

Change-Id: I2bc695688aecf51fe52943ca1e2537ea402d8909