ENH: Add custom tooltip labels to vtkChartXY / vtkPlot
The string vtkPlot::TooltipLabelFormat can be set to provide
user-defined tooltip labels. Full user documentation is provided
in the accessor/mutator declaration of TooltipLabelFormat.

If TooltipLabelFormat is empty (it is by default), a default
tooltip will be shown, identical to the existing functionality.

Internally, tooltip labels now are generated by each vtkPlot itself,
rather than by the containing vtkChartXY, as it was previously.

The protected function vtkPlot::GetCustomTooltipLabel() can be easily
expanded to add additional format flags in the future.

New plot types subclassing vtkPlot may want to override
GetDefaultTooltipLabel() and GetCustomTooltipLabel(), as is done
by vtkPlotBar. It is not recommended that GetTooltipLabel() be
overridden directly.

Change-Id: I9ddcd7b6ee64825681626fe8f4e46b3ef500c9dd