Added k-core decomposition filter to Infovis.
The k-core decomposition is a graph partitioning strategy that is useful for
analyzing the structure of large networks. A k-core of a graph G is a maximal
connected subgraph of G in which all vertices have degree at least k.  The k-core
membership for each vertex of the input graph is found on the vertex data of the
output graph as an array named 'KCoreDecompositionNumbers' by default.