BUG: Set renderer's RenderWindow to NULL in RemoveRenderer.
Without this fix, Renderer::GetRenderWindow() returns a dangling pointer
after vtkRendererer::RemoveRenderer() is called.  This patch also modifies
the destructors of the RenderWindow subclasses so that they all use the
same code block to call SetRenderWindow(NULL) on all renderers, and adds
a check to the vtkRenderWindow destructor to ensure that subclass has
removed the dangling pointer.

Change-Id: Icbf35cb2e6628b67202620d3dcf562d9f0b48e7d