ENH: Dual mode implementation for AMR Reader
This commit introduces a dual mode of operation
for demand-driven AMR readers. The two modes are
as follows:

(1) The first mode employs static load-balancing
implemented on the reader. Downstream modules
send a single request for which blocks to load.
This request is the same on all processes. The
reader then assigns blocks to processes using
block-cyclic distribution.

(2) The second mode loads all the blocks specified
in the request. Downstream modules, send a "different"
request for the blocks that each process requires.
In this case the reader does not re-assign blocks
to processes. Instead it loads the set of blocks
the process requested.