BUGFIX:Employ AMR metadata to determine spacing
Fixed issue with miscalculating the grid spacing to
be used on the resulting grid. Because the input AMR
dataset is not complete -- it is just the set of blocks
required by each process -- figuring out the grid spacing
at level L by looking at the block at level 0 is not a
valid operation since the block at (L,0) may not exist.
Instead, the metadata are used to look up the desired
grid spacing.