ENH: Distance constrained to a polygonal surface mesh
Test demonstrates a distance widget constrained to lie on the
surface of a polygonal mesh. Both handles are constrained to the mesh'
surface. Optionally, one can also specify a height offset. If specified,
the end points of the distance widget are constrained to lie at a height
offset from the surface of the mesh. The "height" at any location on
the surface is measured as the offset of the point in the direction of
the surface normal.