BUG: Trivial producer was not handling extents properly. Fixed.
Trivial producer should not have returned the original dataset if
the requested extent is less than whole and if EXACT_EXTENT() is on.
Because in this case, the pipeline calls Crop() on the dataset later,
corrupting the original dataset for good. Instead, it should create
a copy and pass that to the output. Fixed.