ENH: Add a  point placer that allows one to snap to cell centers.
vtkCellCentersPointPlacer is a class to snap points on the center of cells.
The class has 3 modes. In the ParametricCenter mode, it snaps points
to the parametric center of the cell (see vtkCell). In CellPointsMean
mode, points are snapped to the mean of the points in the cell.
In 'None' mode, no snapping is performed. The computed world position
 is the picked position within the cell. The test demonstrates the
use of the class to constrain handle widgets to the centers of cells. This test
constrains a distance widget to the center of various cell types. However,
it may be used by any of the widgets that use the handles (distance, angle etc).