FIX: Canny Tcl Test. Required Update when passing data to next filter
The method Update needs to be called on a filter if we use GetOutput on that filter and provide the output data object to another filter. Calling Update makes sure that the data object is available. This is only required if we pass data using the SetInputData method.

For example,
 - filterA->Update()
 - filterB->SetInputData( filterA->GetOutput() )