Redesigned the way we read and store parts of the LSDynaReader.
Made the reading of point properties for each part far more efficent
both in terms of time and memory size. Also the vtkLSDynaPart
stores all of the information directly on the unstructured grid allowing
for really nice caching and reduction of deletes/allocs on each timestep.

I also now use ghostcells to hide cells this means that the topology
is constant over the entire adaption level so we can cache that
the entire time.