Fix ParaView cross compilation.
ParaView now calls FindVTK in a couple of places which includes VTKConfig.cmake
long after it includes ParaView3CompileTools.cmake. This replaces
the carefully specified build helper executables that are built in the host
tools build phase, but are not buildable in the cross compilation build phase.

Ideally the two config files should be unified.

Change-Id: I4d6c1f1ffb8a6a0729d23c1a62a9a8cab70b0585