Fixed multiple bugs found during testing.
If a node had no parts but we had properties selected to read it would
attempt to add properties to undefined parts.

The logic to disable InternalEngery cell array for shells was incorrect. We forgot the fact that the parameters passed into a macro aren't checked during the precompiler pass.