Fixed more bugs found during testing.
Correct errors when trying to add to a point property we don't have anything more else to read.

Corrected reading when reading a 64bit file on a 32bit machine to skip
to the right places when reading in parallel.

Simplified the api to get the raw buffer from the reader to only be
the templated version.

Change-Id: Ifc7b8ee8384799a369420dce19fb17c082ec51ad