vtkFixedSizeHandleRepresentation - A marker that has the same size in pixels.
This class is a concrete implementation of vtkHandleRepresentation. It is
meant to be used as a representation for vtkHandleWidget. Unlike the other
represenations, this can maintain a constant size in pixels, regardless of
the camera zoom parameters. The size in pixels may be set via
SetHandleSizeInPixels. This representation renders the markers as spherical
blobs in 3D space with the width as specified above, defaults to 10 pixels.
The handles will have the same size in pixels, give or take a certain
tolerance, as specified by SetHandleSizeToleranceInPixels. The tolerance
defaults to half a pixel. PointPlacers may be used to specify constraints on
the placement of markers. For instance a vtkPolygonalSurfacePointPlacer
will constrain placement of these spherical handles to a surface mesh.
  Also add a test for the same