BUG: Fix track-pad scrolling action for OS X Cocoa.
The Apple track pad generates scrollWheel events for a two-fingered drag.
However, the track pad generates many more of these events than a real
scroll wheel, but for most events the "delta" is zero and only every or
so event has a non-zero delta, this emulates the "click click click"
scrolling of a typical scroll wheel.  VTK was treating the zero delta
values as negative deltas, instead of ignoring them.

Change-Id: I4030ad9fbbf675d5298c3cf4322e489a127e5e26