ENH: Cleaned-up the AMR Resample Filter
1. Renamed VOICHanged to MinMaxChanged to better reflect this is set
when the Min/Max values have been set <--- This is probably not needed
since the ModifiedTime should deal with this.  The default value was
changed to true since when the object is initialy created the values
should be considered changed

2. Created new member data (GridMin, GridMax, GridNumberOfSamples) to
avoid the instance changing the user specified (Min, Max, and
NumberOfSamples).  This prevents the filter from re-executing if the
application sets the same values between updates.

3. ComputeAndAdjustRegionParameters is always called using the metadata

4. Changed SnapBounds not to force the grid to include index (0,0,0) if
it didn't need to.

5. Did some general clean-up

Change-Id: Iea9d56672d28367e36ea7b7f1e8115986939a63b