BUGFIX: Corrected GetCellDims for 1-D and 2-D data
Fixed method to set the cell dims to a minimum of 1 s.t.
when we compute the number of cells by taking the product
of the number of cells in each dimension the result would not
be erroneously zero. This is a temporary fix. The correct solution
is to acquire the data-description and compute the cell dims accordingly.

Change-Id: Ibd37ee9daa17337684958ecc1df2ea9ad235ec20