BUG: Check if input updated since the image mapper updated.
When an input to vtkImageSliceMapper had multiple consumers, if
any of the consumers required a different extent than the mapper,
then the mapper would get the wrong extent on its next update and
crash.  Comparing the input's update time with the mapper's update
time allows the update to be forced when necessary.

Change-Id: I248ccf2b8cb7bd37d0e9ba894775d741c3c389bb