ENH: Initialize ghosted grid points arrays
Modified initialization of ghost data to also initialize the ghosted
grid points arrays. Further, renamed the method "InitializeGhostedFieldData"
to "InitializeGhostData" to better convey the role of the method -- The
method initializes all ghost data both fields and coordinates if the grid
has registered coordinates, i.e., the grid is a structured or rectilinear
grid.

Change-Id: I2f025ac065a40538d889a2fd00434d298ff0f7a8