Add support for scalar opacity when scalars are no being mapped.
When scalar coloring, if the mapper decides that it can directly render the
input scalar array as RGBA, or LA etc., then the mapper should use the opacity
specified in the scalar array for rendering. However, that was not the case. The
mappper simply ignored the scalar opacity.

This patch fixes the issue when not mapping scalars. It's possible that when
mapping scalars, the LUT would return a range for alpha. That case is not
handled by this.