Fix the condition to determine if colors have alpha < 1.
The condition to check if we can avoid passing alpha to the graphics card was
incorrect. Previous commit broke traditional opacity with scalar coloring, while
the code before that didn't respect alpha when scalars weren't mapped. The
right fix is to look at the alpha value range to determine if the colors are
truly opaque.