Added a SelectionMode variable to vtkChart and Scatter Plot
This is a part of the work to add more selection features to the vtk charts.
Five types of seleciton mode can be set to the scene SELECTION_NONE, SELECTION_DEFAULT, SELECTION_ADDITION,  SELECTION_SUBTRACTION,  SELECTION_TOGGLE, which will be used later by the charts in the scene to do these type of selections.

Change-Id: I4abca2d1d19b40859f42e784d97aec95ed278863