BUGFIX: Proper allocation of cell visibility array
Added method to compute the cell dimensions of the structured grid
that are used to compute the total number of cells and to allocate the
cell visiblity array. Previously, the code was erroneously allocating
cell arrays based on the node dimension which was creating unescessary
larger arrays. Also, made the visibility arrays to be allocated when
a visibility array is requested.

Change-Id: I9fea0c4939a9db91df0f48db44ecf09862e3f3c8