Add VTK-specific local commit hooks
Add ".hooks-config.bash" to tell the main hooks to chain to our hooks.
Add a pre-commit hook that cooperates with SetupForDevelopment.sh to
ensure the user has run the most recent hook setup script.  Add a
prepare-commit-msg hook to mention the "WIP:" prefix in the standard
commented commit message instructions.

Change-Id: I034afb64aa4fbbcaafd70d8a57493495d8736181