vtkMPIOutputWindow: Remove unnecessary export modifier
The class vtkMPIOutputWindow had the VTK_PARALLEL_EXPORT modifier
but the declaration of the class is in the cxx and not used elsewhere.
The constructor was also updated to remove unecessary CreateInstance(...) and
delete. There seems to be no reason for this logic and in the case of
a registered overide will blow the stack.

Change-Id: I9af45d5de9a9c74347f1ffb2fa7822fed1c237d5