Making the parametric interior check tolerance consistent.
With the changes to the interior check and cell gradient calculations,
there were some inconsistencies with respect to the tolerance that
was used for determining if a point was contained in a cell
depending on whether or not it was at the apex or not.  This was changed
to be consistent with the original interior check tolerance of 0.001, in
parametric space.  Also, this same tolerance is used to figure out if the
cell gradient calculation parametric point is "near enough" to the
apex point for that specialized computation.

Change-Id: I2ba4098bd54a4b6939a39c6c700bd0a4dc9c6acb