BUG: Fix the wrapping of QVTKInteractor.
In Python, the QVTKInteractor class wasn't inheriting any methods
from its superclass vtkRenderWindowInteractor and was useless as
a result.  The python wrappers were confused because the prefix
is "Q" instead of "vtk".  The main kits have a vtkKitHierarchy.txt
file that the wrappers can use to ascertain type and inheritance
without relying on prefixes, a simple solution here is to use the
hierarchy file that was generated for Rendering.

Change-Id: Idba365ea61d7153dfd63f678dd36318ba429dc93