Fix for MFC compatibility issue with bad_alloc
To be able to catch allocation failures, when VTK is running in the context of other framework (e.g. MFC), we should use (...) in the catch handler instead of std::bad_alloc. If hosting framework implements its own operator new, it's possible that it will throw own defined exception and not the standard std::bad_alloc! In the case of MFC this is CMemoryException, so it would not be handled correctly and crash.

BUG ID 0013018: http://vtk.org/Bug/view.php?id=13018

Change-Id: Ic47b1666116a4299d85e9b86a5d1b13a72822a94