Factor module glob/load into vtk_module_glob macro
Move the module.cmake glob and load logic into a helper macro and invoke
that from the original location.  Parameterize the macro by the VTK
source and build tree to make it re-useable in an external build of the
module test directories.  Pass the list of enabled test languages as
optional arguments.

Change-Id: I3575f73ff759fc096d1df6b2cbe8ff957650f369