Create project to build VTK module tests as an application
Add a "VTKTestExternal" project that builds all VTK module test
directories as a single application outside of the main VTK build.
It may build against a VTK build tree or a VTK install tree.  This
will ensure that VTK presents in its public interface everything that
we test, thus ensuring that everything is available to applications.

Change-Id: I1f46ad5408529eb99d292682a046d4ea659b8033