COMP: Trying to fix java parallel build hiccups in modular.
This is a follow-up to pre-modular commit 875f8232, which was a
temporary fix to the java parallel build failure.  When the build
fails, errors of the form "class file contains wrong class" are
seen.  This patch reorganizes and comments the cmake code for
clarity.  It will probably not fix the problem, but it should at
least assist in determining the cause.

This commit also puts vtkBuildAllDriver.class back in java/vtk,
the modularization commit 0c1471f1d moved it from the java/vtk
dir to the java dir.

Change-Id: Icee402497f2d64a4dd6314a2f31e05c4e0805b91