Produce a more human-readable summary of modules
Mainly sorting lists alphabetically to enable developers to quickly scan
enabled modules (and their dependencies). Producing counts of number of
enabled modules, and the number of dependencies. When the dependency
list is greater than one the dependencies are printed one per line.

Thanks to Chris Harris for working on the summary format with me.

Change-Id: Idb64884628a1712da83aae715d243280f9beb397