Add support for polygonal selection for charts
This adds support for polygonal selection on VTK charts. The
new SELECT_POLYGON mouse action allows for the user to draw
a polygon and when the mouse is released the points contained
inside the polygon are selected.

Currently only the points contained within the convex hull of
the selection polygon are selected.

Change-Id: I23ad29f65e0e9f0b46b09fa8cf678c2d934b6e3a