COMP: Allow wrapping of headers that aren't listed in SRCS.
In pre-modular VTK, there used to be a Kit_WRAP_HEADERS variable
that listed config headers and other headers that did not contain
any class definitions, but which included important constants that
we wanted to wrap in python.

When Kit_WRAP_HEADERS was removed, these headers had to be included
in Module_SRCS in order to be wrapped by python, but then certain
tricks were needed to exclude them from e.g. java.  This commit
moves these headers back to HDRS, and then passes the HDRS to the
various wrappers and allows them to decide what to do with them.

Change-Id: Ieef0a7a42ab7ac0248f3635c6b460e5c08a418d4