COMP: Fix python wrapping of GUISupport/Qt classes in release.
A modification to GUISupport/Qt/CMakeLists.txt in commit 7aaf1d8e,
meant to enhance the wrapping QVTKInteractor, had some serious side
effects on the wrapping in this directory.  These were undetected
because this directory is only wrapped if VTK_WRAP_PYTHON_SIP is ON.
This commit reverts the changes of commit 7aaf1d8e and instead fixes
the QVTKInteractor issue by hard-coding its name into vtkWrapPython.c.
After this is merged, the hard-coded class name will be removed from
vtkWrapPython.c in the master branch.

Change-Id: I1dd2f6227997e9df62710e76459361de4c7b2459