Fix bug when using hyperjump on ScatterPlotMatrix
This fixes a bug in the vtkScatterPlotMatrix class which occurs when
the user clicks outside the valid plot range in hyperjump mode.
Before the fix the scatter plot matrix would jump to the lower-left
plot (0, 0) if the user clicked outside the valid plot area. With the
fix the plot matrix will do nothing for invalid clicks.

Change-Id: Ic5bcd0f4b2b810931e36c4ef0aa20c2f4b4b5c7f