BUG: Fix backwards-compatibility of generated wrapper code.
Comparison of the generated wrapper code to that of a few commits
ago turned up some differences.  Function pointers were not being
wrapped for the Java and Tcl wrappers, and several VTK_SOMETHING
constants were missing from the Python wrappers.  This commit fixes
these issues.  A diff between the new generated code and the old
generated code was used to verify that exactly the same methods
are wrapped as before.

Change-Id: Iae42286824ef6643b42eedfd427763a40be426d1