Auto-init modules that define/implement/use abstract interfaces
The per-module export headers (<mod>Module.h) inject auto-init logic
into applications to ensure that implementations of abstract interfaces
are registered.  Limit invocations of VTK_AUTOINIT to modules that
define or implement an abstract interface.  Include the headers from
each module's dependencies in its own header to propagate auto-init.
This is necessary when a module's implementation uses an abstract
interface from another module not directly used by the application to
ensure that an implementation of said interface is registered.

Change-Id: Ic6ad697f5c0bb3326117bf23bad9b60c40cb3d37