Add code to module infrastruture to export module class information
External wrappers such a ParaView need information about classes
contain in a paricular module. This information was provided by
the vtk${Kit}.cmake files. This information needs to be written to
a new set of files in the VTK_MODULE_DIR along with the appropriate
new macro the load this information.

Change-Id: Iec18d3a03056feb2db5bd54d6c5380c0592c52cf