Move Cosmo classes into vtkFiltersCosmo module
The third party module vtkCosmo can be build with or without
MPI. Rather than building a MPI and non MPI version we still
need the ability to switch between the two implementations. This
can use done by move the Cosmo class into a single module and
providing VTK_COSMO_USE_MPI. This break with rule of not having
a module change it behaviour based on complie options but unless
the third party library is refactored we will have to live with this.

Change-Id: Id36de477d52d1d9b69cc4f2baeabbbdac00c8aaa