vtkXML*: Optionally support 64-bit binary data headers
Teach vtkXMLWriter a "HeaderType" option specifying whether to use
UInt32 or UInt64 as the header element type in binary data within the
output VTK XML file format.  Store the decision in a new "header_type"
attribute of the top-level VTKFile element.  When the type is UInt64 old
readers will not understand the file, so bump the version number from
"0.1" to "1.0".  Default to UInt32 and version 0.1 to write compatible
files when possible.  If an array is too large to be indexed by UInt32
then fail with an error message suggesting use of UInt64 as HeaderType.

Teach vtkXMLDataParser to read binary data headers with the proper
element type.

Change-Id: I7187f70bfc7051f56a4734c8280b7c39c7806370